chore: reorganize package.json
Every npm script was changed into a yarn script. Additionally, a new
script was created in the scripts directory. It does the same as yarn
start, but does so silently.
I chose this way so I can call run-server.sh when running nodemon, while
also allowing yarn start (and the other yarn scripts) to have a verbose
output.