revert: change shell back to sh
The previous change was motivated because of the escape sequences in the
error messages. Although the zsh is the most appropriate to deal with
escape sequences, and is a quite popular shell, not everyone has it
installed in their system. So the scripts are run in sh, and the error
messages are sent to stderr from another script written in CoffeeScript.